+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 <!-- SPDX-FileCopyrightText: Copyright (c) M. Boerger, the MBO Works authors -->
 <!-- SPDX-License-Identifier: Apache-2.0 -->
 
+# 0.3.5
+
 # 0.3.4
 
 - Restore the four directly downloadable platform binaries and additionally publish each platform's